Blacksmith item modifications V1.0 (naming) (#2673)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
Cleaned version of
https://github.com/Rotwood-Vale/Ratwood-Keep/pull/2666 Not sure what
happened...
Allows the blacksmith to name items once they reach full durability.
(made into a list prompt for future possible expansion)
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
It allows people to have custom names for their items and weapons. (any
that go onto an anvil)
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)
https://github.com/Rotwood-Vale/Ratwood-Keep/pull/2666
<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->