Our Sacred Grove Is Being Desecrated (#2558)
## IMPORTANT
This is tagged mapping but it is not. Originally, I had replaced the
wise tree in the grove, but I removed that map change and it's just a
structure which is to be added to the map by mappers.

## About The Pull Request
This PR intends to add important structures to the world which are
sacred. When destroyed they trigger unique bad omens like when you kill
the priest, which have unique events.

-changes the Elderbaum (Elder Tree) from a renamed wise tree to it's own
structure. It is now 5 times mightier than your average wise tree due to
it's important nature.

-when the elderbaum is destroyed, an omen is added to the bad omens list
"OMEN_DESECRATE_DENDOR" and event which uses this specific omen which
triggers an event 2 times and then removes the omen, stopping the event
forever. Each time it triggers, in 5 locations around the map, it will
spawn one of the following:
50% chance to spawn an evil tree that will start spawning creeping vines
25% chance to spawn 3 rous
15% chance to spawn 3 volfs
10% chance to spawn a dryad

-when the elderbaum is destroyed, dendorites will be informed before
everyone else and gain nausea as their equilibrium with nature is thrown
off.

-put's in place the code structure for future shrine desecrations and
specific omen events

## Why It's Good For The Game

Gives a purpose to the druids. Adds hard targets for antags.

## Proof of Testing (Required)

![Screenshot_20](https://github.com/user-attachments/assets/3d5e4f96-1d4f-400f-9dd4-775af50cde04)

![Screenshot_21](https://github.com/user-attachments/assets/d73b2b5d-0b13-4c59-a4ca-e8818592fd57)

![Screenshot_22](https://github.com/user-attachments/assets/87e2c9fb-179a-4e13-8c1e-8eca11c1ae45)