Fixes Fluvian race sprite colors (#2681)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
Makes their sprites more in line with what the others are for the color
conventions
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
It's hard to get the color you want on account that their grey scale
values were all broken. This fixes it.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)
Image of Fluvian more aligned with the color selected in the color
option:

![image](https://github.com/user-attachments/assets/699f3b55-8892-4395-bc9e-0dbce3e4998a)

<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->