Adds Minstrel towner role. (#2615)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
People like to play a musician but currently the only way is to be a
refugee. This allows them to spawn in as a towner. Plus towners need
some love so I might try to put more roles in for them in future.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
The alternative is being a foreigner and traveling a bog to get to town
just to play songs and RP. This will allow a little more music around
town while also allowing those who want to be a minstrel to do so at
roundstart rather than wait for the fog to clear.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)


https://github.com/user-attachments/assets/3fcb9f53-e948-4f79-8234-a3d2bc8df616


<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->