makes all equipped procs keep args of parent proc... (#2688)
...utility added for pickpocketing.

<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
What this does
- Adds a `silent` var for all `put_in_hands` and `put_in_active_hand`
procs
- Standardizes all the `equipped` procs to properly reflect the parent
proc.
- Adds a `silent` arg (defaults to false) to determine if a pickup
should be silent
- adds this silent arg utility to pickpocketing, making the action
silent

No runtimes that I know of. I did not change the procs that overrode the
mob value to be more specific and silent is a named type that only gets
directly called in one instance so it should always default to `FALSE`.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
This not only might prevent future issues with runtime errors from bad
arguments for proc arguments but also allows for expanding silent
pickups for future objects. I've already have others suggest sneaking
could make use of it among other things.

However I primarily did this for the use case of pickpocketing.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)


https://github.com/user-attachments/assets/6061161c-772b-40a4-92c1-861257615d05


<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->