Makes spectacles fixable. (#2692)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
Fixes what https://github.com/Rotwood-Vale/Ratwood-Keep/pull/2679 tried
to fix and failed to. In fact that last Pr probably needs to be un test
merged because I think this was the only issue caused by non 0 object
integrity being broken.

This makes it so when glasses are stepped on they hit 0 object integrity
allowing them to be repaired
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
You cannot currently repair spectacles and there's only 1-2 others n the
map you can use.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)


https://github.com/user-attachments/assets/3aee7191-fec5-4571-a709-c2eb29dc3cf3


<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->