Makes removing bandages require adjacent. (#2709)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
You have to be next to someone, or on their tile, to remove bandages
now.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Another episode of: "Sanity checks? What's that?"
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)
too far

![image](https://github.com/user-attachments/assets/201120ba-2cb3-4610-b3b4-011942b453ef)

Adjacent

![image](https://github.com/user-attachments/assets/d508f120-aef5-4435-8cca-55982a6f4753)

Same tile

![image](https://github.com/user-attachments/assets/43514e1e-3b3f-4a05-b453-c8fa3193992f)

<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->