Adds the blacksmith mod menu. (#2696)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
This should work now.

- Makes the menu for blacksmith mods into a proc
- uses item right clicking to determine to name things.
- Checks if it's at full health and on an anvil to name it.

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
People want it, and it will be expanded in time to allow modifying an
assortment of items.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)

![image](https://github.com/user-attachments/assets/bb6fb005-2d66-4157-9073-e12b86808c85)

<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->