Fixes the infinite fuel bug. (#2698)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
You can spam click a single stick and keep pressing yes to get more
fuel. This does the following:

- Drops the item first (potentially preventing other bugs)
- Does a check if you have anything in your active hand.

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Pretty bad bug to have...
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)


https://github.com/user-attachments/assets/81a0a834-fe8b-41cb-a802-3a5272de52e0


<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->