Adds sound to message spell. (#2699)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
Adds a sound that plays locally to the person receiving the message.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Even with the color it's still hard to catch a message at times. This
might help make it a little clearer, and with time people might help
associate the sound with message. Currently this sound is only used if
someone uses a needle to miraculously heal a wound while having no
skills so it shouldn't be an issue of recognition.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)


https://github.com/user-attachments/assets/48cff9d5-9c98-4761-899a-59f340fda40a


<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->