Physician roles, Acolyte, Hospitaller, Revival, Derot Overhaul (#2662)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
This PR is meant to handle the current issues regarding physicians,
acolytes, pestrans, revivals, and rot. It contains the following
contains the following:

- Farming and climbing removed from towner physician.
- Removes all 'spells' from ALL physician roles with the exception of
secular diagnosis.
- Removes 'De-rot' surgery.
- Adds recipes for 'embalming fluid' used to reset the rot timer when
given to someone (Essence of honey and essence of earth). Toxic to
living beings.
- Physician self aid kits given embalming fluid by default, replacing
medical book.

- Replaces CPR spell with a revival surgery. Ingredient to revive called
'lux' has to be harvested from living humans, (including deadites), and
applied during a surgery to a corpse to revive. Harvesting lux causes
stat penalties to those harvested from and effectively removes one of
that players revives. (Borrowed from Azure peak and modified)

- Adds 'Hospitaller' role. A dedicated pestran role, like Dendor's
druids and necra's morticians. Hospitallers have higher treatment then
acolytes, less holy skill and much less wrestling. They do however, have
apprentice polearms. (Hopefully better sprites for their outfit
pending.)

- Removes Cheele and Pestra's needle from surgical kits. Holy mother of
power creep.
- Court physician granted it's needle of pestra once more in it's bag, 2
lux's in their chambers.
- Tweaks Eoran spells: Now actually cost devotion to use, and the Eoran
Charm no longer hard stuns. Only pacification.
- Removes Miracle from Pestran. 

- Grants Pestrans Pestra's Mercy: A spell that requires 'channeling'
next to the target. Target is given no-die trait as long as miracle is
channeled, and oxyloss is slowly removed from target, with a minimum of
blood added. Channeling the miracle slowly drains devotion. Other
actions can not be taken while in use.

- Reworks Lesser miracle: Lesser Miracle now works as an 'over time'
heal. Affects by various 'conditions' dependant on patron (Borrowed from
Azure peak and modified).
- Removes Miracle from Astratans

- Grants Astratans Smite Unfaithful: A miracle that activates instantly
on screen wide effects. Targets all excommunicated and apostates,
setting them on fire, providing a chance to stun, and granting a debuff.
This spell is twice as strong in the day time, and during day time it
will affect the undead including vampires, deadites, and skeletons.

- Tweaks cure rot: Curing rot now leaves a trait behind called 'rot
touched' on those cured. Those with rot-touched will have a low chance
for scarred wounds to open every 15 minutes, signifying the consequence
of being afflicted by the rot plague. Further more, examining a rot
touched person grants a significant mood penalty to the viewer.

- Tweaks rejuv elixor: No longer heals all wounds, Instead heals
infected wounds, Now actually metabolizes in a body.

- On request of administration: Rewrites Mori's code for hypospray stys
and bottles to be less atrocious. My god was this painfully bad. It
should be much better to administrate, as well as actually function the
way it was likely intended.

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Our current situation has been an odd one to say the least. Towner
physicians, despite being an unlimited towner role proved to be better
at healing and reviving then the church due to it's ability to revive,
derot and heal. This has the unfortunate consequence of ruining a great
deal of ratwood's themeing, with an apocalyptic plague.

So, We've cut back Physicians from being able to rapid heal people via
spells. They're still able to revive, but have to seek resources to do
so. Unfortunately, Farming had to be removed from physicians, as
physicians proved to not be using it for alchemy. They'll have to get
farmers to grow things for them, or learn it the slow way. Physicians
have lost the ability to de-rot, in exchange for gaining a way to delay
rotting via applying embalming fluid. This keeps our thematic setting of
the rot being something thats consuming the world, and can be delayed,
but only cured through divine intervention.

Speaking of that. Pestrans recieving the same miracle as an astratan
didn't sit well with me in the slightest, So we've replaced it with a
life saving miracle, able to keep someone alive and pull them out of a
dangerious point at a cost of steadily draining their devotion. However,
due to themeing of the server, Cure rot was made imperfect.
While I am debating on whether the occasional bleed, and people looking
at you getting a harsh penalty is too much of a punishment, It should
help with fixing the server themeing, and as I said with physicians,
rotting can now be delayed.

Similarly in regards to themeing, Astratan revives no longer cause
severe stat penalties. However, They can only use anastasis on a person
/once/. If that same person returns dead, they will be unable to be
revived by anastasis.
Astratan's have also had their miracle replaced with 'smite unfaithful'
a churn undead similar spell used on excommunicated and apostates.
During the day it also affects undead.

Hospitallers: I hate to add a new role. Especially as maintainers are in
the process of cutting roles from the server. But at the moment, it's
somewhat needed. Due to us allowing differant patron types for acolyte,
It has lead to acolytes often not having a pestran or an astratan. An
astratan can be excused, as the priest is an astratan, but lacking a
pestran becomes a critical gameplay issue. The alternative to adding a
dedicated patron role, is re-restricting acolyte patrons; and I believe
that is probably a rather unpopular change to the community. The cat is
out of the proverbial bag, and putting it back in would be unpleasant.

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)

![image](https://github.com/user-attachments/assets/255c4602-e638-41ea-a92e-39a1e629e33c)

![image](https://github.com/user-attachments/assets/bfbf3485-4f05-4709-9b9d-d8c0daf6e296)

![image](https://github.com/user-attachments/assets/ca7ade2e-38b6-453c-a1a5-fbbb81029edc)

![image](https://github.com/user-attachments/assets/9b6ed4d4-5ace-4ae1-a7d8-d75f32ce28e1)

![image](https://github.com/user-attachments/assets/92619522-e44c-42a8-9bc0-5e471c92de5c)


![image](https://github.com/user-attachments/assets/9b5caf56-3268-409d-8e9b-5f48be94da30)

![image](https://github.com/user-attachments/assets/99d701dd-ab7f-495f-8820-7e1c94f32baf)

![image](https://github.com/user-attachments/assets/ab494462-aa05-4da6-aaaa-4ecdc1bd1175)

![image](https://github.com/user-attachments/assets/0c93c438-8512-434e-b52e-2fd858b75a79)

![image](https://github.com/user-attachments/assets/37ebdf27-5bdd-4227-b926-fe0098190ecf)

<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->

---------

Co-authored-by: Gyrandola <pasta.frollagg@gmail.com>
Co-authored-by: Sutures <101909198+noelle-lavenza@users.noreply.github.com>