Templar helmet selection (#2728)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
This PR adjusts Templar setup logic to remove the random helmet
selection for Astratan and Necran Templars, replacing it with a menu
offering a choice between the two on round start.

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
This PR was requested by a friend who sometimes plays an Astratan
Templar. From looking at the code, and at the items in-game, there's no
mechanical between the two helmets aside from one having a visor that
can be lifted. Lifting the visor offers no combat advantage (and in fact
disadvantages the wearer by reducing its protection), so the randomness
appears not to have been a balance decision.

As such, it seems harmless to give players the option to choose a helmet
according to their preference. In particular, I'm aware that at least
some admins have, in the past, been willing to swap out the helmet in
response to an ahelp - this will obviate that process, saving both
player and admin time.

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)
The menu:

![image](https://github.com/user-attachments/assets/3b28309d-1073-4d81-8261-0ffbda134464)

Helmet slot after queueing as an Astratan Templar and choosing the
visored helmet:

![image](https://github.com/user-attachments/assets/602bccd1-c721-4b92-b535-f7102e4771cb)

Helmet slot after queueing as an Astratan Templar and choosing the
bucket helmet:

![image](https://github.com/user-attachments/assets/85d6ab3a-2796-4760-beb9-31f9241b1717)

Helmet slot after queueing as a Necran Templar and choosing the visored
helmet:

![image](https://github.com/user-attachments/assets/f35c98e4-66df-4719-aaa4-6bbeb951744f)

Helmet slot after queueing as a Necran Templar and choosing the bucket
helmet:

![image](https://github.com/user-attachments/assets/6cfed961-b68b-4499-9f0a-7e4d5943bf2d)

I've also ensured that loading as Templars of other gods does _not_
create the menu - though if appropriate helmet choices are added for
them the logic could easily be modified.

<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->

## Addendum
I'm not really that familiar with the proper terms for mediaeval and
renaissance arms and armour, so if there would be a better way to refer
to the visored helmet than as "Visored" in the menu (and likewise for
the bucket helm), let me know and the string can be changed.