Makes feints work against dodge (#2650)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

Makes feinting work against dodge. Also slightly buffs feinting while on
feint intent to make the immobilize 2 seconds instead of 1.2 seconds

## Why It's Good For The Game

Dodge based classes are notoriously bad to try and fight. They usually
have more stamina than you and cannot be grabbed or attacked as they
have a 90% chance to just dodge it. Now feints work exactly the same for
if they are parrying or dodging


## Proof of Testing (Required)

It was a single line change, nothing broke when tested.  
Proof it compiles

![image](https://github.com/user-attachments/assets/fd2cbd90-df35-4084-9720-8a8651a40934)