Expands goldface access. (#2602)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
Allows anyone who can read, and has access to an unlocked goldface, to
access the secret menus.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
It's a pain for a solo shop hand to run the merchant if the merchant
dies or is away. It would also make it more fun if people could also
mess with the menus for their own benefits. Of course they'd still need
a key. it might also make merchant slightly more exciting if people find
out they could access certain things this way without your explicit
permission. Up to you. If you want I can just edit it to include shop
hands but I think anyone being able to access it works well.

Not sure if this breaks anything else but the portable gem device seems
to be unused so it should be fine?

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)


https://github.com/user-attachments/assets/6039f434-c7c9-4cfa-8d7e-a877513c8e60



<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->