Caveling Shaman Buff + Fix (#2749)
## About The Pull Request

Gives Underdweller Shaman a few skills, as well as a magical tome
equivalent to the Battle Magos and fixes the hemostat.

- Sewing 0 -> 1 - Shaman might often make clothes or sew fabrics for
rituals, it would make sense for them to have a slight idea of it.
- Cooking 0 -> 1 - As above, Shaman may also cook or make basic meals
for rituals or celebrations.
- Miracles 0 -> 1 - Despite having Churchling miracles, Shaman
apparently have no idea what they're doing with them, or seemingly how
they're doing it.
- Reading 0 >- 1 - The Shaman are capable of miracles and magic, yet are
completely unable to read. They should have a basic idea, and this
allows them to actually learn some magic.
- Apprentice Spell Book - The Battle Magos gets this, and has identical
Arcane and Reading to the Shaman, I don't think it's too strong to give
it to them.

Oddly, the Shaman was supposed to get two primitive clamps, but it was
coded incorrectly. This fixes such.

If any of these changes too strong, I am willing to change them. 

## Why It's Good For The Game

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

The Underdweller Shaman is a flavorful holy/magic hybrid, and somehow
fails at both. As of now, it doesn't have a book or holy skill, which is
impressive that they're healing people with a miracle and casting magic
without having much of an idea of what it is.

The class is racelocked to Vermin and Kobolds, and as such they have
pitiful strength- as they should, but they fall incredibly short
compared to many other classes. This class is seemingly designed for
support and utility, and has been given two skills to match what a
Shamen might actually do while giving them the ability to learn magic
and know what a miracle is.

## Proof of Testing (Required)


![image_2025-04-18_170054206](https://github.com/user-attachments/assets/9ffbece2-b10b-46e4-9480-6d344d8639e4)

![image_2025-04-18_170058656](https://github.com/user-attachments/assets/141d5a68-537f-4f62-8065-562bd5b5955a)

Co-authored-by: Envoy <31294289+cDiggidyDawg@users.noreply.github.com>