Update soilson.dm (#2742)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
This is a minor QOL change for soilsons that gives them a pouch, a
dagger and a satchel.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
I believe it's very annoying and unreasonable for soilsons to spawn
without a pouch, satchel or proper belt. They start with 1 storage slot,
in their rope belt. I think these changes are reasonable.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)
I simulated some tests in my head and it seems ok.
<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->