Fix for initial memories missing character titles (#2703)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

This PR modifies initialisation logic for characters so that their
initial memories - and others' memories of them - are only set after
their role is fully initialised, meaning that the memories a player
begins with will now correctly include their entire "real name" and
role.

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

Prior to this PR, initial memories would include the name of a
character, but not their title. For most roles, this is, at worst, a
mild inconvence, but there are a couple of situations where it's more
significant:
- Spellcasters who learn certain spells, such as _Message_, need to know
the real name (including title) of a character to correctly cast the
spell on them. I think this works even if they aren't in your memories
as long as you can provide the title, but I could be wrong about that.
[edit: I've since been informed that you _do_ need an accurate name in
your memory for _Message_ to work)
- Court Magi can only scry a person by giving a correct real name _and_
by having that real name recorded in their memories - the erroneous
names given on round start are _not_ adequate for this.

Consequently, this PR is largely a significant quality-of-life
improvement for court magi, who no longer have to dart around the Keep
on round start in order to examine everyone they might later be asked to
scry, a smaller quality-of-life improvement for other spellcasters, and
a relatively insignificant bug fix for everybody else.

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)

Testing methodology: on a test server, I (Bronden) and one tester
(James) joined at round start. I readied as Court Magos and the tester
readied as Duke (after I awarded some PQ to them). Each of us verified
our memories were correct and I scried the Duke. After these steps, a
second tester (Brandeth) late-joined as a Knight. We again each
collected screenshots of our memories to verify that they were correct
and I scried the Knight.

Lobby showing myself and another player on a test server, myself as
Court Magos

![image](https://github.com/user-attachments/assets/b224f422-ba2a-43d7-b6e0-49dd10634acc)

My initial memories on round start, _without_ examining anybody:

![image](https://github.com/user-attachments/assets/8077adef-6e2a-42bb-9c94-874860c3459f)

Corresponding initial memories of "Duke James":

![image](https://github.com/user-attachments/assets/d04eb0ac-674c-4d65-aee7-a70188fad2db)

Screenshot of myself scrying "Duke James", again without any prior
examination:

![image](https://github.com/user-attachments/assets/0ed4131c-f648-4850-87cb-33c58516dbce)

Screenshot showing my memories after a second tester late-joined the
server as a knight:

![image](https://github.com/user-attachments/assets/8f4ba475-447f-4b64-9d5f-4f51fee97512)

Screenshot showing the late-joiner's initial memories, once more without
examining anybody:

![image](https://github.com/user-attachments/assets/32efe7b7-40c8-4427-8ab6-73660c5891a9)

The first tester's memories after the late-joiner joined the game (the
red text on the right is simply a PQ warning from initial attempts to
ready as Duke):

![image](https://github.com/user-attachments/assets/46eed8f3-1ae2-4aae-8bd5-3b5ba548d387)

And finally, a screenshot of myself scrying the late-joined Knight,
again without any examination having taken place:

![image](https://github.com/user-attachments/assets/52ddb9f6-7073-4bf9-992b-603a479c9da3)

As you can see, each character's real name was correct in all cases.


<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->