Languages from SS13 and aren't used needs tobe removed (#2753)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

Removing all languages from SS13 that aren't used in Roguetown

Removed the following languages:

- Drone 
- Encoded Audio Language (Machine)
- Chimpanzee 
- Mushroom
- Nar'Sian
- Piratespeak
- Slime
- Swarmer
- Xenomorph
- Beachbum

## Why It's Good For The Game

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

Makes code cleaner and aids admins and eventmins in need of giving
certain language and not bombarded with a ton of languages we do not
use.

## Proof of Testing (Required)

<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->

[It's really just code-cleaning, all proof needed can be found in the
lines removed]